Move incoming sync parsers to database utilities
